refactor: extract shape/gradient/shadow code into typst_gen_shapes module
Move 8 shape-related functions (generate_shape, write_shadow_shape,
write_fill_color, write_shape_params, write_shape_stroke,
write_polygon_vertices, write_polygon, write_gradient_fill) from
typst_gen.rs into a dedicated typst_gen_shapes.rs submodule.

This reduces typst_gen.rs from 1307 to 1050 lines with no logic changes.
The new module follows the same pattern as typst_gen_tables.rs and
typst_gen_diagrams.rs.
